Added tests for location controller

+++ server/controllers/locations.js
@@ -1,5 +1,6 @@
 const locationRouter = require("express").Router();
 const Location = require("../models/Location");
+const Item = require("../models/Item");
 
 locationRouter.get("/:id", (req, res, next) => {
     const id = req.params.id;
@@ -7,7 +8,7 @@ locationRouter.get("/:id", (req, res, next) => {
     Location
         .findById(id)
         .then(result => {
-            if (result) return res.json(result);
+            if (result) return res.json(result.toJSON());
             else res.status(404).end();
         })
         .catch(next);
@@ -16,29 +17,54 @@ locationRouter.get("/:id", (req, res, next) => {
 locationRouter.get("/", (req, res, next) => {
     Location
         .find({})
-        .then(result => void res.json(result))
+        .then(result => void res.json(result.map(r => r.toJSON())))
         .catch(next);
 });
 
 locationRouter.post("/", (req, res, next) => {
-    const body = req.body;
-    const locationName = body.name;
+    if (!req.authenticated) return next(new Error("UNAUTHORIZED"));
+    if (!req.authenticated.staff) return next(new Error("FORBIDDEN"));
 
-    if (!locationName) return res.status(400).json({ error: "name is missing" });
+    const { name } = req.body;
+
+    if (!name) return res.status(400).json({ error: "name is missing" });
 
     const newLocation = new Location({
-        name: locationName
+        name
     });
     newLocation
         .save()
         .then(result => {
-            res.status(201).json(result);
+            res.status(201).json(result.toJSON());
         })
         .catch(next);
 });
 
-locationRouter.delete("/:id", (req, res, next) => {
+locationRouter.put("/:id", (req, res, next) => {
+    if (!req.authenticated) return next(new Error("UNAUTHORIZED"));
+    if (!req.authenticated.staff) return next(new Error("FORBIDDEN"));
+
+    const { id } = req.params;
+    const { name } = req.body;
+
+    if (!name) return res.status(400).json({ error: "name is missing" });
+
+    Location
+        .findByIdAndUpdate(id, { name }, { new: true })
+        .then(result => void res.json(result.toJSON()))
+        .catch(next);
+});
+
+locationRouter.delete("/:id", async (req, res, next) => {
+    if (!req.authenticated) return next(new Error("UNAUTHORIZED"));
+    if (!req.authenticated.staff) return next(new Error("FORBIDDEN"));
+
     const id = req.params.id;
+
+    const itemUsingThisLocation = await Item.findOne({ location: id });
+
+    // TODO: Check the http status code
+    if (itemUsingThisLocation) return res.status(409).json({ error: "there are items using this location" });
 
     Location
         .findByIdAndRemove(id)
